Added MetastoreTypeInfo similar to TypeInfo
This patch adds classes similar to TypeInfo called MetastoreTypeInfo in standalone-metastore.
Ideally, we should move TypeInfo to standalone-metastore since they store the information
about types. However, moving TypeInfo to standalone-metastore is non-trivial effort primarily
because of the below reasons:

1. TypeInfo is annotated as Public API.
This means we can only alter/move these classes in a compatible way.
2. Directly moving these classes is not straight-forward because TypeInfo uses PrimitiveEntry
class which internally maps the TypeInfo to Type implementations. Ideally metastore should
not use Type implementation which makes it harder to move the TypeInfo directly.

However, if we are ready to break compatibility, then TypeInfo broken such that it doesn't
use PrimitiveEntry directly. In such a world TypeInfo will store just what it needs to store.
Metadata of Types i.e the type category, its qualified name, whether its a parameterized type
or not and if yes, how do we validate the parameters.

I am assuming that breaking TypeInfo is a no-go and hence I am copying the relevant code
from TypeInfo to Metastore and calling it MetastoreTypeInfo. MetastoreTypeInfo and its sub-classes
are used by TypeInfoParser (also copied) to parse the column type strings into TypeInfos.