Moved TypeInfo and sub-classes to standalone-metastore
This patch moves TypeInfo and its sub-classes to standalone-metastore. This means that hive-serde module needs to depend on standalone-metastore and the
pom.xml of hive-serde adds a new dependency to standalone-metastore. In addition to moving the TypeInfos we also need to move some Util classes like
Category, PrimitiveCategory, PrimitiveEntry to standalone-metastore.

In TypeInfoFactory the default TypeInfos for various primitives are declared as static finals and are accessed all over the place. This is not ideal
because char, varchar and decimal uses attributes from the Type implementations which cannot be accessed by standalone-metastore. I had to remove
these 3 TypeInfos and move them to PrimitiveObjectInspectorFactory instead.