[OPTIONAL] remove static references of TypeInfos from TypeInfoFactory
The TypeInfoFactory class has static final references to various PrimitiveTypeInfos which are used
by Hive code at many places. For non-parameterized TypeInfos it is fine, but in case of parameterized
TypeInfos like varchar, char and decimal this is a problem since when we move TypeInfoFactory to
standalone-metastore, we cannot depend on the type implementations. For consistency reasons this
change moves these TypeInfo constants to TypeInfoUtils from TypeInfoFactory