Polling based approach.
This is a alternative proposal. The Driver is issued
a client identifier when the URL is returned.
The subsequent call from the Driver passes the client id.
This call is blocked until HS2 knows the validation result.

Some thoughts:
1. We could implement a small timeout based poll so that we
don't hold up resources on server for unauthenticated clients.
2. The race conditions (validation received but client has not
polled yet and/or client coming in before validation) need to be
solid. The driver doesn't really know how long its going to take
for the SSO flow so having a good timeout value is tricky.
We can always make it configurable. The trade-off here is that
we are consuming server resources (CPU cycles and a thrift thread)
until validation is done.
3. A malicious user can just keep sending these polls without
any validation ever receiving and server will block until the
time-out which doesn't sound like a great thing about this.